Enum quill_mir::mir::StatementKind

pub enum StatementKind {

    Assign {
        target: LocalVariableName,
        source: Rvalue,
    },
    InstanceSymbol {
        name: QualifiedName,
        type_variables: Vec<Type>,
        target: LocalVariableName,
    },
    Apply {
        argument: Box<Rvalue>,
        function: Box<Rvalue>,
        target: LocalVariableName,
        return_type: Type,
        argument_type: Type,
    },
    InvokeFunction {
        name: QualifiedName,
        type_variables: Vec<Type>,
        target: LocalVariableName,
        arguments: Vec<Rvalue>,
    },
    ConstructFunctionObject {
        name: QualifiedName,
        type_variables: Vec<Type>,
        target: LocalVariableName,
        curry_steps: Vec<u64>,
        curried_arguments: Vec<Rvalue>,
    },
    InvokeFunctionObject {
        func_object: Rvalue,
        target: LocalVariableName,
        additional_arguments: Vec<Rvalue>,
        return_type: Type,
        additional_argument_types: Vec<Type>,
    },
    StorageLive(LocalVariableId),
    StorageDead(LocalVariableId),
    DropIfAlive {
        variable: LocalVariableName,
    },
    Drop {
        variable: LocalVariableName,
    },
    Free {
        variable: LocalVariableName,
    },
    ConstructData {
        ty: Type,
        variant: Option<String>,
        fields: HashMap<String, Rvalue>,
        target: LocalVariableName,
    },
    ConstructImpl {
        aspect: QualifiedName,
        type_variables: Vec<Type>,
        definitions: HashMap<String, LocalVariableName>,
        target: LocalVariableName,
    },

}

Variants

Assign

Moves an rvalue into a local variable.

Fields of Assign

target: LocalVariableNamesource: Rvalue

InstanceSymbol

Creates a local instance of a definition such as a function (or in some cases, a constant). This statement is removed by the “func_objects” pass, where curried functions are removed.

Fields of InstanceSymbol

name: QualifiedNametype_variables: Vec<Type>target: LocalVariableName

Apply

Applies one argument to a function, and stores the result in a variable. This statement is removed by the “func_objects” pass, where curried functions are removed.

Fields of Apply

argument: Box<Rvalue>function: Box<Rvalue>target: LocalVariableNamereturn_type: Typeargument_type: Type

InvokeFunction

Invokes a function with the given arguments. This is inserted by the “func_objects” pass. MIR generation should instead try to generate InstanceSymbol and Apply statements, and this pass will convert them (where necessary) into this statement.

Fields of InvokeFunction

name: QualifiedNametype_variables: Vec<Type>target: LocalVariableNamearguments: Vec<Rvalue>

ConstructFunctionObject

Creates a function object which contains a function and any curried arguments. It may be later invoked by InvokeFunctionObject to execute the function with the provided arguments. curry_steps states the amount of arguments that are required in each step to execute the function object. In the func_objects pass, curry_steps is a list of n ones where n is the arity of the function. In later passes, the curry steps might be changed - but they will always sum to n. For example, a function (a, b, c) -> d with curry_steps = [1, 1, 1] will be represented as a chain of function objects a -> (b -> (c -> d)). With curry_steps = [2, 1], the function will be represented as a chain of function objects (a, b) -> (c -> d). This is inserted by the “func_objects” pass.

Fields of ConstructFunctionObject

name: QualifiedNametype_variables: Vec<Type>target: LocalVariableNamecurry_steps: Vec<u64>curried_arguments: Vec<Rvalue>

InvokeFunctionObject

Invokes a function object with the arguments it contains, along with perhaps some additional arguments. This is inserted by the “func_objects” pass. The amount of additional_arguments must exactly match the number of arguments remaining in the function object.

Fields of InvokeFunctionObject

func_object: Rvaluetarget: LocalVariableNameadditional_arguments: Vec<Rvalue>return_type: Typeadditional_argument_types: Vec<Type>

StorageLive(LocalVariableId)

Hints to LLVM that this variable’s lifetime has now begun, and that we may use this variable later. This is automatically generated by the borrow checker.

StorageDead(LocalVariableId)

Hints to LLVM that we will no longer use this variable. This is automatically generated by the borrow checker.

DropIfAlive

Calls the destructor for the object, and deallocates the variable’s memory, if it has a destructor and it is currently alive. This is automatically inserted by the MIR generator at the end of the scope declaring a variable. Please note that a drop-if-alive instruction in MIR does not necessarily call the actual destructor, only if the object really is not moved out. In particular, the first MIR drop-if-alive instruction an object encounters will perform the actual drop and deallocation, any more are no-ops and will be optimised away.

Fields of DropIfAlive

variable: LocalVariableName

Drop

Calls the destructor for the object, if it has a destructor. This is automatically generated by the borrow checker. This does not deallocate the memory containing this object.

Fields of Drop

variable: LocalVariableName

Free

Deallocates the memory containing this variable. Similar to a drop, but does not run the destructor code. Use this when a data structure is being destructured, meaning that we cannot call the destructor. In particular, if this variable is on the heap, this results in a free call, and if on the stack this is a no-op. This is automatically generated by the borrow checker.

Fields of Free

variable: LocalVariableName

ConstructData

Creates an object of a given type, and puts it in target.

Fields of ConstructData

ty: Typevariant: Option<String>

If this type was an enum, which variant should we create?

fields: HashMap<String, Rvalue>target: LocalVariableName

ConstructImpl

Creates an impl of an aspect from a set of definitions. The definitions are instanced symbols. They are considered to be “moved into” the impl for ownership purposes.

Fields of ConstructImpl

aspect: QualifiedNametype_variables: Vec<Type>definitions: HashMap<String, LocalVariableName>target: LocalVariableName