Enum quill_mir::mir::StatementKind

pub enum StatementKind {
    Assign {
        target: LocalVariableName,
        source: Rvalue,
    },
    AssignPhi {
        target: LocalVariableName,
        phi_cases: Vec<(BasicBlockId, LocalVariableName)>,
    },
    InstanceSymbol {
        name: QualifiedName,
        type_variables: Vec<Type>,
        special_case_arguments: Vec<KnownValue>,
        target: LocalVariableName,
    },
    Apply {
        argument: Box<Rvalue>,
        function: Box<Rvalue>,
        target: LocalVariableName,
    },
    InvokeFunction {
        name: QualifiedName,
        type_variables: Vec<Type>,
        special_case_arguments: Vec<KnownValue>,
        target: LocalVariableName,
        arguments: Vec<Rvalue>,
    },
    ConstructFunctionObject {
        name: QualifiedName,
        type_variables: Vec<Type>,
        special_case_arguments: Vec<KnownValue>,
        target: LocalVariableName,
        curry_steps: Vec<u64>,
        curried_arguments: Vec<Rvalue>,
    },
    InvokeFunctionObject {
        func_object: Rvalue,
        target: LocalVariableName,
        additional_arguments: Vec<Rvalue>,
    },
    Drop {
        variable: LocalVariableName,
    },
    Free {
        variable: LocalVariableName,
    },
    ConstructData {
        name: QualifiedName,
        type_variables: Vec<Type>,
        variant: Option<String>,
        fields: BTreeMap<String, Rvalue>,
        target: LocalVariableName,
    },
    ConstructImpl {
        aspect: QualifiedName,
        type_variables: Vec<Type>,
        definitions: BTreeMap<String, LocalVariableName>,
        target: LocalVariableName,
    },
}

In MIR, the program is in static single assignment form: each variable is assigned once only. A variable called target is typically where we store the result of a statement.

Variants

Assign

Fields

target: LocalVariableName

source: Rvalue

Moves an rvalue into a local variable.

AssignPhi

Fields

target: LocalVariableName

phi_cases: Vec<(BasicBlockId, LocalVariableName)>

A list of source blocks and the variable that we should move into target if this is where we came from. In the case that we came from a different block, the given variable will not have been initialised.

Moves an rvalue into a local variable conditionally, depending on which block we just came from.

InstanceSymbol

Fields

name: QualifiedName

type_variables: Vec<Type>

special_case_arguments: Vec<KnownValue>

target: LocalVariableName

Creates a local instance of a definition such as a function (or in some cases, a constant). This statement is removed by the “func_objects” pass, where curried functions are removed.

Apply

Fields

argument: Box<Rvalue>

function: Box<Rvalue>

target: LocalVariableName

Applies one argument to a function, and stores the result in a variable. This statement is removed by the “func_objects” pass, where curried functions are removed.

InvokeFunction

Fields

name: QualifiedName

type_variables: Vec<Type>

special_case_arguments: Vec<KnownValue>

target: LocalVariableName

arguments: Vec<Rvalue>

Invokes a function with the given arguments. This is inserted by the “func_objects” pass. MIR generation should instead try to generate InstanceSymbol and Apply statements, and this pass will convert them (where necessary) into this statement.

ConstructFunctionObject

Fields

name: QualifiedName

type_variables: Vec<Type>

special_case_arguments: Vec<KnownValue>

target: LocalVariableName

curry_steps: Vec<u64>

curried_arguments: Vec<Rvalue>

Creates a function object which contains a function and any curried arguments. It may be later invoked by InvokeFunctionObject to execute the function with the provided arguments. curry_steps states the amount of arguments that are required in each step to execute the function object. In the func_objects pass, curry_steps is a list of n ones where n is the arity of the function. In later passes, the curry steps might be changed - but they will always sum to n. For example, a function (a, b, c) -> d with curry_steps = [1, 1, 1] will be represented as a chain of function objects a -> (b -> (c -> d)). With curry_steps = [2, 1], the function will be represented as a chain of function objects (a, b) -> (c -> d). This is inserted by the “func_objects” pass.

InvokeFunctionObject

Fields

func_object: Rvalue

target: LocalVariableName

additional_arguments: Vec<Rvalue>

Invokes a function object with the arguments it contains, along with perhaps some additional arguments. This is inserted by the “func_objects” pass. The amount of additional_arguments must exactly match the number of arguments remaining in the function object.

Drop

Fields

variable: LocalVariableName

Calls the destructor for the object, if it has a destructor. This is automatically generated by the borrow checker. This does not deallocate the memory containing this object.

Free

Fields

variable: LocalVariableName

Deallocates the memory containing this variable. Similar to a drop, but does not run the destructor code. Use this when a data structure is being destructured, meaning that we cannot call the destructor. In particular, if this variable is on the heap, this results in a free call, and if on the stack this is a no-op. This is automatically generated by the borrow checker.

ConstructData

Fields

name: QualifiedName

type_variables: Vec<Type>

variant: Option<String>

If this type was an enum, which variant should we create?

fields: BTreeMap<String, Rvalue>

target: LocalVariableName

Creates an object of a given type, and puts it in target.

ConstructImpl

Fields

aspect: QualifiedName

type_variables: Vec<Type>

definitions: BTreeMap<String, LocalVariableName>

target: LocalVariableName

Creates an impl of an aspect from a set of definitions. The definitions are instanced symbols. They are considered to be “moved into” the impl for ownership purposes.